temp_Cam/requirements.txt
flask
flask-cors
paho-mqtt
websockets